Refactored code to use common namespace registration code
    -- moved common code out of origin_ui into server_ui
    -- Adjusted cache_serve and origin_serve
