+++ server_ui/register_namespace_test.go
@@ -16,7 +16,7 @@
  *
  ***************************************************************/
 
-package origin_ui
+package server_ui
 
 import (
 	"crypto/ecdsa"
